Move shared WASM logic to base writer
